Move the TerminalColors type to ansi_codes
This places it closer to related code.

Bug: b/507940443
