+++ base/cvd/cuttlefish/ansi_codes/BUILD.bazel
@@ -12,3 +12,12 @@ cf_cc_library(
         "@abseil-cpp//absl/strings",
     ],
 )
+
+cf_cc_library(
+    name = "terminal_colors",
+    srcs = ["terminal_colors.cc"],
+    hdrs = ["terminal_colors.h"],
+    deps = [
+        "//cuttlefish/ansi_codes",
+    ],
+)
