+++ base/cvd/cuttlefish/ansi_codes/terminal_colors.cc
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "cuttlefish/ansi_codes/terminal_colors.h"
+
+#include <string_view>
+
+#include "cuttlefish/ansi_codes/ansi_codes.h"
+
+namespace cuttlefish {
+
+std::string_view TerminalColors::Reset() const {
+  return is_tty_ ? kAnsiReset : "";
+}
+
+std::string_view TerminalColors::BoldRed() const {
+  return is_tty_ ? kAnsiRed : "";
+}
+
+std::string_view TerminalColors::Red() const { return is_tty_ ? kAnsiRed : ""; }
+
+std::string_view TerminalColors::Cyan() const {
+  return is_tty_ ? kAnsiCyan : "";
+}
+
+}  // namespace cuttlefish
