+++ base/cvd/cuttlefish/ansi_codes/terminal_colors.h
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include <string_view>
+
+namespace cuttlefish {
+
+class TerminalColors {
+ public:
+  TerminalColors(bool is_tty) : is_tty_(is_tty) {}
+  std::string_view Reset() const;
+  std::string_view BoldRed() const;
+  std::string_view Red() const;
+  std::string_view Cyan() const;
+
+ private:
+  bool is_tty_;
+};
+
+}  // namespace cuttlefish
