+++ base/cvd/cuttlefish/host/commands/cvd/BUILD.bazel
@@ -30,14 +30,14 @@ cf_cc_binary(
     srcs = ["main.cc"],
     deps = [
         ":libcvd",
+        "//cuttlefish/ansi_codes:terminal_colors",
         "//cuttlefish/common/libs/utils:environment",
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:subprocess",
         "//cuttlefish/common/libs/utils:tee_logging",
         "//cuttlefish/flag_parser",
         "//cuttlefish/host/commands/cvd/cli:log_files",
-        "//cuttlefish/host/commands/cvd/cli:utils",
-        "//cuttlefish/host/commands/cvd/utils",
+        "//cuttlefish/host/commands/cvd/utils:common",
         "//cuttlefish/host/commands/cvd/version",
         "//cuttlefish/host/libs/vm_manager",
         "//cuttlefish/posix:strerror",
